adds a script to run the whole pipeline; adds result cleaning
