+++ conversation_reconstruction/dataflow/conversation_reconstruction/setup.py
@@ -101,7 +101,8 @@ REQUIRED_PACKAGES = [
     'google-cloud-storage == 1.6.0',
     'google-apitools == 0.5.22',
     'NoAho==0.9.6.1',
-    'yamlconf==0.2.3', 
+    'mwparserfromhell==0.5.1',
+    'yamlconf==0.2.3',
     'mwtypes==0.3.0',
     'beautifulsoup4==4.5.1']
